+++ lib/src/rtc_data_channel.dart
@@ -67,6 +67,7 @@ abstract class RTCDataChannel {
   Function(RTCDataChannelState state)? onDataChannelState;
   Function(RTCDataChannelMessage data)? onMessage;
   Function(int currentAmount, int changedAmount)? onBufferedAmountChange;
+  Function(int currentAmount)? onBufferedAmountLow;
 
   /// Get current state.
   RTCDataChannelState? get state;
@@ -77,7 +78,10 @@ abstract class RTCDataChannel {
   /// Get channel label
   String? get label;
 
-  int get bufferedAmount;
+  int? get bufferedAmount;
+
+  /// Set threshold to trigger onBufferedAmountLow callback
+  int? bufferedAmountLowThreshold;
 
   /// Stream of state change events. Emits the new state on change.
   /// Closes when the [RTCDataChannel] is closed.